+++ src/express/routers/combineRouter.js
@@ -1,7 +1,7 @@
 const express = require('express');
-const KeywordLists = require('../models/keywordLists');
-const TotalInformations = require('../models/totalInformations');
-const common = require('../express/common');
+const KeywordLists = require('../../models/keywordLists');
+const TotalInformations = require('../../models/totalInformations');
+const common = require('../routes/common');
 
 const router = new express.Router();
 common(router, KeywordLists, '/keywordList');
