+++ src/express/routes/deleteUser.js
@@ -1,4 +1,4 @@
-const auth = require('../middleware/auth');
+const auth = require('../../middleware/auth');
 
 module.exports = (app, queries) => {
   return app.delete(queries, auth, async (req, res) => {
